Fix: 뱃지의 requirement 필드를 Long 타입으로 변경

+++ greenmap/src/main/java/com/ucamp/greenmap/badge/service/BadgeServiceImpl.java
@@ -31,7 +31,7 @@ public class BadgeServiceImpl implements BadgeService {
         return BadgeResponse.builder()
                 .name(badge.getBadgeName())
                 .wholePoint(point.getWholePoint())
-                .currentPoint((long) badge.getRequirement())
+                .currentPoint(badge.getRequirement())
                 .description(badge.getDescription())
                 .image_url(badge.getImage().getImageUrl())
                 .created_at(badge.getCreatedAt())
